requirements.txt
fastmcp==2.12.4
httpx>=0.27.2
pydantic>=2.7.1
python-dotenv>=1.0.1
pytest>=8.2.0
fastapi>=0.111.0
uvicorn>=0.30.0
openai>=1.35.0
langchain>=0.2.0
pyyaml>=6.0.1
